requirements.txt
requests
schedule
beautifulsoup4
selenium
python-slugify
python-gnupg